refactor(whispering): move key bindings out of Tauri
Make utils/key-binding.ts:9 the frontend shortcut vocabulary owner now that Rust no longer reads structured bindings. Remove the compatibility re-export and its one-caller file.
