+++ apps/whispering/src/lib/tauri/commands.ts
@@ -114,10 +114,6 @@ export type {
 	TranscriptionError,
 	TranscriptionSpec,
 } from './bindings.gen';
-// The shortcut-binding vocabulary is FE-owned now that Rust no longer sees a
-// structured binding (ADR-0117); re-export it here so consumers keep importing
-// from `$lib/tauri/commands`.
-export type { Key, KeyBinding, Modifier } from './key-binding-types';
 /**
  * Typed event listeners generated by tauri-specta. Each leaf
  * (`events.dictationCapabilityEvent`) binds the topic name to its payload type,
