+++ apps/whispering/src/lib/utils/reserved-shortcuts.ts
@@ -12,8 +12,11 @@
  *   (ADR-0117).
  */
 
-import type { Modifier } from '$lib/tauri/commands';
-import { type BindingLike, isRegistrableChord } from '$lib/utils/key-binding';
+import {
+	type BindingLike,
+	isRegistrableChord,
+	type Modifier,
+} from '$lib/utils/key-binding';
 
 /**
  * `primary` stands for the platform's command modifier: Command on macOS,
