+++ apps/whispering/src/routes/(app)/(config)/settings/shortcuts/keyboard-shortcut-recorder/create-chord-recorder.ts
@@ -1,7 +1,12 @@
 import { debounce } from '@epicenter/workspace';
 import { on } from 'svelte/events';
-import type { Key, KeyBinding, Modifier } from '$lib/tauri/commands';
-import { domCodeToKey, eventModifiers } from '$lib/utils/key-binding';
+import {
+	domCodeToKey,
+	eventModifiers,
+	type Key,
+	type KeyBinding,
+	type Modifier,
+} from '$lib/utils/key-binding';
 
 const CAPTURE_WINDOW_MS = 300; // Time to wait for additional keys in a combination.
